Remove RegionSwitcher update guard
Previously, the RegionSwitcher would only update when the player moved
to a new block. I originally did this to reduce workload, but if the
overhead were actually an issue then the player would be stalling every
time they move to a new block, which is equally unacceptable. I'd rather
have the overhead of checking every tick, since it makes changes kick in
immediately.